--- src/client/scripts/js/gridView.js
@@ -1,9 +1,0 @@
-var gridView;
-(function (gridView) {
-    class MyExport {
-        constructor() {
-            this.myVal = "hello from file gridView";
-        }
-    }
-    gridView.MyExport = MyExport;
-})(gridView || (gridView = {}));
