--- src/client/scripts/js/view/Rounds.js
@@ -1,5 +1,0 @@
-class Rounds {
-    constructor() {
-        this.roundOne = roundOne;
-    }
-}
