--- src/client/scripts/ts/view/Rounds.ts
@@ -1,4 +1,0 @@
-/// <reference path="rounds/RoundOne.ts"/>
-class Rounds{
-  roundOne: number[][] = roundOne;
-}
