--- src/client/scripts/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-        "target": "es6",
-        "removeComments": true,
-        "outDir": "js"
-    },
-  "files": [
-    "ts/app.ts"
-  ]
-}
